Move `createAction` test helpers to top level of repo.
