+++ common/types.d.ts
@@ -21,8 +21,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-import type { Steps } from '@elastic/synthetics';
+import type { ActionInContext, Steps } from '@elastic/synthetics';
 
+export type ActionContext = ActionInContext & { isOpen?: boolean };
 export type StepStatus = 'succeeded' | 'failed' | 'skipped';
 export type JourneyType = 'suite' | 'inline';
