+++ src/hooks/useDrop.test.ts
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-import { createSteps } from '../helpers/test';
+import { createSteps } from '../../common/helper';
 import { computeIsDroppable } from './useDrop';
 
 describe('useDrop', () => {
